requirements.txt
SQLAlchemy
asyncpg
aiohttp
Werkzeug
pydantic==1.*
pydantic[email]
python-dotenv
isort

